Add order component added and api handling
Add order component added and api handling
